Add a "diagnose" button to test result server
This adds a simple endpoint which downloads the test logs from S3 and runs them
through the parse_test_failure script.  I renamed this script and did a small
refactor so I could easily import the code from the test result server.

The diagnose button is a bit slow, but this quick hack is still useful.  At
some point we should get this pre-calculated in the database, but when I tried
altering the table to add a blob column, all of the queries got slow (too
bad mysql is not columnar!)

Change-Id: Idcf26d68c6a8bec63442ab4fa3bff45a25bf93c4
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5209
Tested-by: jenkins
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>