Bump minimum required cmake version to match thirdparty
If you accidentally run cmake with an older version (eg using the default one
in the $PATH on older Ubuntus) you can end up with hard-to-diagnose build
issues. This sets our minimum required version to match what we ship in
thirdparty, which forces people on older distros to use that version.

Change-Id: I224200b6f7e119d12222ca8b28b14942cbcc83c7
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5102
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>