Change log indexing to be based on REPLICATE ranges
This changes the footer of the log to keep a simple min/max of the ids
of replicate messages in the log, removing the half-baked indexing code
that was there previously.

It also changes several log APIs to take only indexes instead of OpIds.
Several call-sites now "bridge" APIs by only passing through the index
of an OpId -- we intend to change over these APIs to be index-based
as well.

This also updates the read path for replicate entries to support replaced
entries. A test is added which generates a randomized sequence of log messages
and rolls and validates that the correct entries are read back, even
if entries have been replaced with different terms.

Change-Id: I1c199004980b426fd27d126342d2bd631f520938
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5103
Tested-by: jenkins
Reviewed-by: David Alves <david.alves@cloudera.com>