Make followers send their last known committed index in ConsensusStatusPB
This makes followers send their committed index in ConsensusStatusPB
when they reply to the leader.

Having this extra information is useful for aborts. If the replica
replies that the log matching property failed, and the leader doesn't
have the replica's last received id (because it was overwritten) then
they still have a index in which they can agree to have in common.
The leader can then just send all operations starting at the replica's
last committed index.

Change-Id: I7712b9d68ee276eb94a6a8cfe2c4497bb702ba47
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5135
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>