Transaction abort - Consensus side
This implements the consensus side of operation abort, including aborting/
overwriting running transactions if a new leader pushes other operations in
their place. This adds an unit test that asserts this works as expected,
i.e that the replica aborts the correct operations.

Follow up patches will include the queue side and an itest that ties it
all together.

Change-Id: I08f37a3af71786215ade0f94c70cc63b5b7fdd36
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5137
Reviewed-by: Michael Percy <mpercy@cloudera.com>
Tested-by: jenkins