Change LogCache::ReadOps() to use an index instead of an OpId
Now that the log is based on indexes we can just change this method
to use indexes (which is actually necessary, when looking up committed
indexes for which we don't know the term).

This also changes the reader/async reader/tests so that, when we read
in an interval, the starting point is inclusive instead of exclusive.

Change-Id: Ib7df8be0474f4bafb5f4e0c10e04da58791121d8
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5149
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>