Transaction abort - Consensus Queue side.
This adds the final component required for transaction abort, i.e. allow
the queue to handle when peers report a last received that does not exist
in the state machine (because it needs to be truncated/overwritten).

This adds a detailed unit test but still doesn't add the itest, the point
being that a proper itest that tests everything together needs some additional
features such as proper stepdown/promotion.

Change-Id: Id3ed9b3bb6b91c3bf142a729e61bf7264f87b02c
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5214
Tested-by: David Alves <david.alves@cloudera.com>
Reviewed-by: Todd Lipcon <todd@cloudera.com>