KUDU-109. log: add checksums for each entry batch, fix corruption detection
This modifies the WAL format so that each entry is prefixed with a header including:
  - length of the data
  - checksum of the data
  - checksum of the above two fields.

This serves to detect corruption or truncation in either the header itself or
in the actual log data.

One common cause of corruption is if the log segment is pre-allocated -- in
that case all or some of the message data would be '\0' bytes. This sort of
corruption is actually allowed -- it happens naturally when a server crashes in
the middle of WAL write.

To detect this "allowed" part of corruption, the reader code now scans forward
from the point of corruption to the end of the file, looking for any further
data which appears to be a valid entry header (ie a header with a
self-consistent checksum). If it finds one, then this indicates that this is a
"true" corruption instead of just a corruption/truncation of the last batch.

In terms of testing, a new unit test is added which flips a byte in a log. I
also locally tried adding a sleep() in between writing the header and the batch
data and ran linked_list-test. Because linked_list-test kills servers, this
often resulted in a failure to restart before, since it would read null bytes
for one of the log entries. Now, it properly handles this case.

This does not fully address KUDU-109 -- we also should add CRCs on the header
and footer.

This should also address one source of the flakiness in linked_list-test (KUDU-361).

Change-Id: I04d71e828dcfc5e94dd4eb9216d43f695e3b139b
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5229
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>