Fix leader election proxy ownership
Currently consensus owns the proxies that are used to communicate with
remote peers in leader election. This is problematic in a few ways:
- More state for consensus to manage according to the current quorum.
- Proxies were being created up front making mocking of consensus hard.
- Weird ownership bugs could arise.

This takes care of that by making VoterState own the proxy and dispose
of it on destruction.

On the way to address this the ThreadPool ownership of the multiple
proxy implementations needed to change, now each proxy is passed a
pool that it can reuse.

The main reason with the proxy can't own the pool is that we would
get this infamous exception on pool destruction:
"Check failed: rv == 0 (35 vs. 0) . Resource deadlock avoided"

Change-Id: Ifb565a4c5f8d0b562ff1e7d79905182a10de594b
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5236
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>