Remove the bind address from master.flags
The multi-master code uses the bind address to tell the TS where the master leader is,
which we unfortunately set to 0.0.0.0 so the TS are all confused.

In the future we shouldn't rely on the bind addresses since it means we can't bind to
the wildcard address and that we need a different config file per master (or have it
passed on the command line for each master, which is what we're going to do).

Change-Id: I51a73dd5933d829d2e4cbc73b376a6d33e05d1ca
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5269
Reviewed-by: Alex Feinberg <afeinberg@cloudera.com>
Tested-by: Jean-Daniel Cryans <jdcryans@cloudera.com>