Allow to select an index to be assigned the leader role in  BuildQuorumPBForTests()
We currently use this to generate the test quorums and we always assign the last
peer the CANDIDATE role. This adds a new argument so that we can select a specific
index to be LEADER (of course, in that case the last peer won't be assigned the
CANDIDATE role anymore).

Change-Id: I61d2a3e6af9a0fc73d36b293a16895cd13f038b4
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5280
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>