Fix bug in LogCache where it would set the wrong preceding id if the searched index was MinimumOpId().index()
The cache works by searching for the operation with a given index and,
on return, setting the id of the op with the searched index as the
'preceding_id'. This works great except when loading operations for
empty peer's whose preceding index is 0 (from MinimumOpId().index()).
Since the operation with index 0 doesn't exist we would instead set the
operation with index 1 as 'preceding_id', which was wrong.

This patch fixes this by making the AsyncLogReader use the searched index
as a parameter for the callback and then having LogCache give special
treatment for when the searched index is 0.

Change-Id: I407e64a27ad85bad2130eee0767fee6f8cead1e2
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5281
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>