Fix bug in client where it wouldn't select a replica other than the leader
We have two distinct failure "levels" at the client, for writes. We have
a "total failure" scenario where the replica is not available and the
"not the leader" scenario where the replica is available just not the leader
of the particular tablet we want to write to.

We were handling the latter wrongly. In particular even though we added the
old leader to the set of followers, if it hadn't failed we would pick it up
again instead of trying another replica.

This patch fixes this by making sure we only pick the leader if we haven't
added it to the followers before.

This also adds another Status for which we use the "not the leader" failover
scenario: Status::Abort(). This is required since this is the status returned
when transactions are aborted because a new leader was elected.

Change-Id: If8f4e5fe4905925e1fcce36bbd3f5312df65d237
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5282
Reviewed-by: David Alves <david.alves@cloudera.com>
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>