Rename /catalog-manager endpoint to /tablez
The string /catalog-manager is too long (messes up the navbar UI) and
also Catalog Manager is kind of an internal concept whereas a Tables
page should be obvious to anyone.

That said, I think we should make our navbar easier to use some time
down the line and not expose the URLs, so we can keep 37337 URLs but use
nice labels for them in the navbar.

Change-Id: I34bd68adac0dbf55c5eec860f3dfa6c7f1f3add5
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5287
Reviewed-by: Alex Feinberg <afeinberg@cloudera.com>
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>