Simplify the wiring of Log into PeerMessageQueue and LogCache
Ever since we've added the ability for the queue to asynchronously
load messages from the log we've needed to plumb the log through it.
We were getting it from the consensus interface but we want to make
consensus a mere observer of the queue regarding watermark advancement,
which should not provide the log.

This just passes the Log to PeerMessageQueue, which builds the LogCache
with it.

Change-Id: Ib40b43f10847c5f9a37581b1725756dc83972c09
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5294
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>