Disable leader election by default
Leader election can sometimes run when there are long pauses such
as in TSAN's linked_list-test. This is the right behavior the problem
is that without the missing consensus features often it can't run
successfully.

This patch disables leader election by default, we can enable it once
the rest of the consensus patches get in.

Change-Id: I786f1dd4882f34a0716539d66d95367391ac0b21
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5296
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>