Fix bug where Random.Next64() produced only 36 bits
Also added a unit test that verifies that the current RNG algorithm (by
design) generates 31 bits of randomness for a 32-bit int and 62 bits of
randomness for a 64-bit int.

TODO: Fix the above to add more randomness. It doesn't look simple to
fix this without changing the algorithm, except by using a 64 bit
integer to do the calculations for the 32-bit return value. But it'd be
worth a closer look later.

Also moved the Random.Normal() unit test from random_util-test to the
new random-test.

Change-Id: If6c42e016e1123d0b6903d77f1d53e7c3b000d4d
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5307
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>