Remove 'last_replicated_' from consensus
We originally came up with the distinction between 'last_received_' and
'last_replicated_' so that we could have multiple requests in flight
and so that we could report to the leader while waiting for the fsync
to finish. However we only still allow for one request to be in-flight
and we've added another lock so that we don't report to the leader until
an Update() is done, otherwise we'de get subtle bugs. This makes the two
watermarks unecessary for now, and coordinating the two is tricky with
peer promotions and step downs.

This patch simplifies the state of consensus a bit by completely removing
the 'last_replicated_' watermark. We can add it back later if we need it,
and when we refactor to have a central watermark tracker or something of
the sort it will be easier to keep them coordinated.

Change-Id: I6acb1af37b813ab76d7747d51f352b4565fec6ce
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5318
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>