Add Close() method to LogCache, make it dump non-persisted ops on close
We close the PeerManager before the queue, which owns the cache. This means
that when we dispose of the cache there might be operations in flight that
were never written to the Log by the local peer. I initially had a fatal
check instead of the dump code and saw that happen semi-frequently.

This just makes the cache dump whatever was not in the log previously.
Since consensus closes the queue and subsequently the cache under its
lock this is safe.

This is kind of an intermediate solution. A proposed refactor is supposed
to unify the log writing paths which would likely also address this, but
in the meanwhile this allows us to make progress.

Change-Id: I9f337ddf3bdc991aa30e87877574ad9907a17fda
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5328
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>