Make consensus observe the queue instead of having the queue explicitely update consensus
Right now we have consensus updating the queue with new operations to replicate
and then we have the queue updating consensus back when the majority replicated
op is updated. This cyclic dep is problematic and error prone.

This patch simplifies the relationship by making consensus a mere observer of
the queue. When consensus is about to shutdown is simply stops observing the
queue.

Change-Id: I04ae8f9f67b5097d4fdbe07eef7c73842e692b61
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5336
Tested-by: David Alves <david.alves@cloudera.com>
Reviewed-by: Todd Lipcon <todd@cloudera.com>