Allow to Close() PeerMessageQueue and have it report Status::IllegalState() when closed
This allows to reuse the queue by resetting the watermarks appropriately and
by closing the cache. This also makes the queue return Status::IllegalState()
when closed instead of crashing: When we become followers the TabletPeer might
still try to replicate operations as if we were leader, we need refuse those
operations but we should not crash.

Change-Id: Ic51ec456bc346a8e27862e0212ee186fde5c215f
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5337
Tested-by: jenkins
Reviewed-by: David Alves <david.alves@cloudera.com>