KUDU-433 - Part1 - Implement and test step down and re-promotion and add an itest that pauses leaders and runs multiple elections
This implements step down and re-promotion by cleaning up state where relevant.
Currently, this means closing the queue and the PeerManager, but a refactor is in
flight where we'll only need to stop the PeerManager and the queue will survive
role changes.

This adds an itest that, while writing with multiple threads, pauses the leader
runs a leader election and then unpauses the leader. This is performed
(2 * kNumReplicas) - 1 times so that we're sure that the same node gets to step
down/re-promote multiple times.

Note that this does not use automated leader election yet. The point being that
I've seen failed election causing problems and won't to leave those out of this
particular test. We may add those as a follow up.

This is a part of series of patched that will mark this issue finished. Missing
still are:
- Using automated elections in the itest.
- Using a mix of kill and stop in the itest.

Change-Id: I2b6741a404a2fb7e4692736e7d920af40cecb276
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5343
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>