Allow not to use EmulateElection at startup time, add exp. backoff to leader election, and make the failover integration test use automated leader elections
This patch adds code to RaftConsensus to support removing
EmulateElection. Currently the code relies on the failure detection flag
to determine which code path will be executed.

This also makes some changes so that leader elction works
on a larger cluster, namely:
- Makes elections exponentially backoff when a peer votes for
  someone else, of when it loses an election. The backoff is
  exponential and is based on the distance between the current
  term and the term of the last committed operation.
- Snoozes the FD when the candidate loses (not only when it
  wins). Not snoozing the timer on lost elections was basically
  making elections run forever.

Finally this refactors the failover test in raft_consensus-itest
to use automated elections.

This patch was co-contributed by David and Mike.

Change-Id: I82c2156932728250eec594daeaa58bee38851b6e
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5333
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins