Improve logging in master
* Log at ERROR level when the catalog manager fails to init.

This would happen anyway in a real master, but I spent an hour debugging
a master_replication-itest failure where the error never got reported.
The test was waiting on the first master to start up, but that master
was blocked on the second, which had failed. After adding this LOG(ERROR)
it became obvious what was wrong.

* Include tablet name and peer in some of the log messages during startup.

Change-Id: I6d6fb8a16bcc79e995b8293c47a5d0c0b90fe4a4
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5357
Tested-by: jenkins
Reviewed-by: David Alves <david.alves@cloudera.com>