Change leader to write to local log via ConsensusQueue
This changes the ConsensusQueue and LogCache to act more like a write-through
cache fronting the log. When the leader proposes new operations, it still
writes them to the ConsensusQueue, but now the ConsensusQueue itself directly
appends them to the local log.

When the local log append is successful, the callback is handled by
ConsensusQueue itself, which generates the "fake" response and updates its
internal watermarks. We plan to refactor this into a separate class in the
future.

The PeerManager no longer tracks a Peer object for the local log. Instead it
only is in charge of RPCs to remote peers.

Because appending to the queue now also appends to the log, we no longer
want to append pending replicates at startup time. If we did that, we'd end
up writing those messages in duplicate to the log unnecessarily. Instead,
we let the LogCache lazily load them back from disk like any other earlier
messages.

Follow-up work:
* Collapse the Peer class hierarchy since it is now only RPC peers. This
  isn't trivial since we have tests which rely on subclassing there.
* Change the replica path to also write via ConsensusQueue, so that the
  ConsensusQueue doesn't need to be torn down and rebuilt on role change.

Change-Id: I940ddf22c8710f1f86e635e33a8c0578e0933f10
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5359
Reviewed-by: David Alves <david.alves@cloudera.com>
Tested-by: jenkins