Master RPC: add master_rpc library, parallelize RPCs.
Extracts Master RPCs common to client and tserver to a separate
master_rpc library (which doesn't depend on master and can
be used in client, tserver, and master).

Makes GetLeaderMasterRPC send requests to all Masters in parallel
from the client: this significantly speeds up client recovery,
especially when one of the masters is stopped, paused, or otherwise
timing out (as opposed to failing 'cleanly', e.g., refusing
connections).

Change-Id: Ib4927d63dde9ef744e55d24973925f69ef98dfa0
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5367
Reviewed-by: Michael Percy <mpercy@cloudera.com>
Tested-by: jenkins