Refactor the queue and fix watermark advancement bug
This refactors the queue to stop using ConsensusStatusPB to track
the peers. We instead now use plain fields in the TrackedPeer struct
instead. This allows to cleanup a bit and also makes our lives
easier when we change to assembling the consensus requests elsewhere.

This also fixes a two bugs (and adds assertions/tests for them), namely:
- Prevents a SIGSEGV when trying to advance a watermark with less peers than
  it requires.
- Makes sure we only advance watermarks based on successful exchanges with
  peers.

Finally this fixes a couple of places in consensus_queue-test where we were
using a local and a remote peer but had a majority of 1 (likely unchanged
from before we automatically appended to the local peer).

Change-Id: If5b59984ac080521732309f14b5cc0356ce8d90a
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5387
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>