Avoid spinning when loading ops for a peer, after an LMP mismatch
On the first round of messages, when a peer replies to the leader with an LMP
mismatch error which forces the leader to read from disk, we are currently
continuously sending and receiving messages until the operation is loaded from
disk.

This is a regression. Our previous mechanism to avoid this was to have the leader
send status-only requests to peers with their true last received operation.
This mechanism caused no LMP mismatch while the queue is loading from disk asynchronously
and thus no spinning. However, on one of the review iterations for the abort patches we
changed this behavior and made the queue report its own last appended instead, until the
operations are loaded from disk. This change causes the replica to always reply
with a LMP mismatch, which always caused 'more_pending' to be set to true and
thus another request to be sent immediately after.

This fixes this by reporting 'more_pending' on LMP mismatch only for the
first round. This will make the leader immediately read from the cache
and, if the cache has the ops the peer needs in memory, immediately update
the peer. However, if the operations need to be loaded from disk, we wait
for a heartbeat or until the next signal that a request is pending.

This adds an assertion to a current unit test that makes sure that this works.

Change-Id: I4f9b7202d7fda45d4db2541bcf4aa4c3a9a5363f
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5388
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>