client: fix generated cmake file to be compatible with older cmake
When trying to build Kudu-Impala integration on el6, I found that I wasn't able
to use the generated cmake module file with the old version of cmake that is
installed by default on that system. Even though we can require Kudu itself to
build with a recent cmake (from thirdparty), we can't impose the same
restriction on consumers of our client.

This patch adds a workaround to the top of our generated cmake file which fixes
the issue by providing the ${CMAKE_CURRENT_LIST_DIR} variable when it is
missing.

Change-Id: I0b87624c494a66c05aa4d8881f8bbd9ee954a1ec
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5711
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins