Make the queue/cache reusable and merge all consensus write paths
This merges all the consensus write paths (except the commits) to
use the queue (and the LogCache underneath) to write to the local
log.

This patch makes it so that we can reuse the queue/cache across role
changes, makes tracking/untracking remote peers atomic and fixes some
bugs that would appear due to reuse.

This patch also makes notifying queue observers asynchronous so that we
completely decouple queue and watchers, which addresses a few deadlock
scenarios.

Finally this also fixes the hack where we were releasing the consensus lock
to close the peers.

Change-Id: I683c0b0a302221d99e225bd6aed33007762d4346
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5433
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins