Reduce duplicated code in KeyEncoder using templates
Although we usually discourage template magic in our style guidelines,
I think this is a net win, since it removes a lot of duplicate code.
In a follow-on patch, I am going to be improving the decoding of partial
row keys, and I didn't want to implement that improvement 8 times.

As a side effect of the code change, now we crash if we try to pass
an invalid (too-short) encoded key slice into KeyEncoder::Decode. This
caused TabletServerTest::TestWebPages to fail where it used to silently
access invalid memory. This patch disables the test: a follow-up patch
will fix the bug.

Change-Id: Ie102af7971253612fee27204b32d97e5f49bd6dd
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5421
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>