Simplify code for key encoding, fix string decoding
This cleans up the code used for encoding string portions of composite keys. It
also fixes a bug where we didn't properly decode escaped nulls.

This adds a TODO where the decoding function should be able to return a bad
Status if the encoded key is invalid, but currently preserves the existing
behavior of crashing.

Change-Id: Id6ddbc58ed2b61a7d700259f4e49f911e6674521
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5486
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>