Fix SIGSEGV in transaction abort
Currently, when we abort transactions due to log truncation in consensus,
we delete the transaction immediately. If the transaction is still preparing
we then get a segmentation fault.

This fixes abort to work more like before the transaction refactoring, i.e.
the transaction is marked for abort, but only actually aborts at synchronization
points.

Change-Id: I86d6e6df11e807e1accae2465175527e46d2a886
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5475
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>