Move *WITH_PREFIX() log macros to their own file and make the queue, cache use them too
This improves the debuggability of the cache and queue by making them also use
the *WITH_PREFIX() statements we were also using in raft_consensus.cc and raft_consensus_state.cc.
This required wiring the tablet_id as well as the peer uuid through the queue and
the cache, but it's a small price to pay as the log statements are much, much, easier
to follow.

To avoid duplicating the macros everywhere they were moved to their own file.

Change-Id: Ie5ba25e9cb35b3408821539d7c26385f0217e17e
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5470
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins