Fix debug-printing of empty encoded keys
We use empty keys to signify the start and end of a table.  Previously we were
relying on some lucky behavior that an empty std::string has a good-sized
buffer of \0 characters in it, and thus we didn't crash when we tried to
DebugEncodedKey(...) the empty keys.

This patch adds special case stringifications for empty start/stop row keys.

Change-Id: I8ae3fa9719a9d6457801419407a9952fe6f631a8
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5487
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>