client: Make speculative TS failover sticky
Currently, the client attempts to try other replicas in a quorum
per-batch, when sending an RPC. However, it doesn't update the meta
cache, so it retries the broken leader each time a batch is sent until
the Master is updated. If the master takes a while to refresh, this can
cause a lot of traffic to the Master and a lot of failed RPCs.

This patch preemptively marks followers we will attempt to treat as the
leader as LEADER in the meta cache, so that future writes will go
straight to that follower. If we turn out to be right in our speculative
guess, and the chosen follower becomes the leader, then we will not have
to go back to the Master for a metadata refresh. If the chosen follower
turns out to be just a follower, then we will mark it as a follower when
we get a failure reply, and thus when we run out of possible leaders we
will refetch the metadata from the Master, as we do today, and the whole
process starts over.

Change-Id: I4b701047ded05334ffee9af46a42a38583cc17fa
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5505
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins