Fix flakyness in RaftConsensusITest::TestAutomaticLeaderElection
This test is currently flaky because we prune the replicas so that we'll test
only live servers for identical data, but the master still reports the old leader
as alive. When we try to get that guy from the current set of replicas we get
a SIGSEGV because it isn't there anymore.

After this fix I ran the test in a loop for 50 iters and couldn't see a failure,
whereas before it would fail in < 10 iters.

Change-Id: Ibff7709e9ff0bf28a815a3d2681033311395c9c5
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5500
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins