log_cache: fix memory accounting when operations are replaced
This fixes a bug where, when a message is replaced in the log cache, we weren't
properly releasing that memory usage from the MemTracker.  This caused a sort
of slow leak of accounted bytes in the memtracker in a long-running cluster
test, eventually causing the log_cache to refuse to accept any more operations,
despite only having a small amount of real memory usage.

Change-Id: Iea42c64b3f4f800afdd21fb8bb4f4680206da46b
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5498
Reviewed-by: David Alves <david.alves@cloudera.com>
Tested-by: jenkins