Fix flakyness in raft_consensus_quorum-test on RaftConsensus::Shutdown()
When we shutdown consensus we close the queue inside the consensus lock.
Since closing the queue implies waiting for all notifications to complete,
this means that we might be waiting for someone to update consensus while
holding the consensus lock.

This just makes it so that we close the queue outside of the lock.

Running:

KUDU_ALLOW_SLOW_TESTS=1 build/latest/raft_consensus_quorum-test --gtest_filter=*TestConsensusStopsIfAMajorityFallsBehind* --gtest_repeat=100
along with:
stress --cpu 10 --io 10 --vm 10 --hdd 10

would hang with a deadlock in 5-10 iterations, while with this fix I was
able to run 100 iterations without it hanging.

Change-Id: Ib99a85a1406e0549feb09171331c5b8f24b051b6
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5501
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>