Bug fix: On RaftConsensus::Update() fill the response before releasing the lock
We currently have a bug where a replica might reply to the leader,
after successfully completing an Update() call, with a term that is
higher than the leader's. This makes the queue crash as the leader is
expecting the replica to reply with INVALID_TERM, if it has a higher
term,

This happens because there is a race in Update() where we only fill
the response after dropping the lock, meaning even if the update
was successful an election might be triggered increasing our term
before we fill the response to the update call. This of course only
happens if the log is slow.

This patch makes it so that we fill the response before dropping the
lock, making sure the response reflects the status of the replica
right after the update was performed.

We should, however, also make sure that we don't trigger the timer
while we're waiting on our own log.

Change-Id: I881f39c502547bcd0dee730b4d1044395b8e1f76
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5504
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins