Make replicate messages ref counted
Also:
Bug Fix: Queue/cache doesn't move the pin backward immediately when messages are overwritten;
Bug Fix: Deadlock between the queue/cache/log on overwrites, when waiting for the Log to sync

In the real cluster we were seeing a rare bug where, on overwrite, a follower
would not move the cache's pin backward. This happened because the queue refused
to update any global watermark (i.e. the all replicated index) when the peer's last
received watermark that changed was lower than the global one we wanted to update.

This fixes it so that we do change the global watermark in that case. This also
adds a test that, without the fix, would trigger the exact failure we've seen
in the cluster.

Addressing this bug also required to change some things inside of log cache itself
and messing with the 'inflight_to_log' operations. Since we also discovered that
the hang we're getting is due to a deadlock while waiting for the log, just so that
we can manage the lifetime of ReplicateMsgs in the cache, the best solution to both
these bugs, which also improves/reduces the code overall and reduces copies, was to
ref count the replicate messages.

This patch thus refcounts all the replicate messages that are shared between the
transactions, consensus, queue, cache and log, making our lifes easier and requiring
less copying around and solving the hang we were seing at the same time.

This only place where replicate messages could still use refcounting would be the peers
which would make the cache completely free, but that should be done in a follow up
patch.

Change-Id: I8d34a9b4228d80daeb426933d6662d1171b4d323
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5509
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>