scan_spec: Replace vector<EncodedKeyRange> object with separate bounds members
Previously, ScanSpec contained a vector<EncodedKeyRange*>, but most of the
usage sites only handled the case of having a single EncodedKeyRange. This
added a lot of unnecessary complexity (as evidenced by the fact that this diff
has a lot of negative lines).

In all cases, if we had multiple EncodedKeyRanges, we could narrow them down by
intersecting them, so we may as well just store the upper/lower bound directly
in ScanSpec.

Change-Id: Id87ab7ede180c17713885ddb25f6f21ae5ab4314
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5515
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins