Make block caching tunable
Currently, when we read a cfile block, we cache it. There are cases where we don't want to do
this like during a compaction, or when issuing a large scan.

This patch makes all the necessary plumbing, so that all the iterators pass down a cache control
enum that tells the cfile reader whether it should cache or not. It was possible to use
ScanSpec in some places, and it had to be added in others. Metrics were also added to differentiate
requests that want to cache blocks VS requests that don't. This is useful when answering questions
like: "how often do we hit cache when we want the blocks to be cached?".

We also introduce two new enums. CacheBehavior in cache.h indicates if we need to increment the
_caching metrics. CacheControl in cfile_reader.h indicates if we should cache the blocks we read or
not, and it has a 1-1 mapping with CacheBehavior. They were separated for two reasons: 1) we don't
want cache.h to leak everywhere and 2) they really mean different things, since one is about metrics
and the other is about caching a block or not.

As of this patch, everything requires block caching except for compactions and only for data blocks.
Bloom files and index files, when read via a compaction, will still cache those blocks. Some command
line tools also disable caching.

The client-side changes will be done in a latter patch.

This patch was tested on a cluster with YCSB but it didn't show us going faster, or slower. It could
be because our caching was already good enough to serve the random reads.

It was also tested on a single machine with 1GB of data. Without the patch, after the load phase,
we end up with about 1GB in the block cache. With the patch, and after all the compactions are
done, there's only ~5MB in the block cache.

Change-Id: Ib40790192d385cf8288bba5c19960c14591b3810
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5206
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>