[java client] Support leader failover
This patch first fixes our handling of a leader failover by recognizing that illegal state
and aborted errors are symptoms of it so that we start poking the other TS in the tablet's
list. We do this by marking them as leader so that we hit them after a retry. After hitting
all the nodes we go back to the master, retry the whole (newish) list, etc.

This patch also adds a helper method to detect if all the errors in a RowsWithErrorException
are of the AlreadyPresent kind, which we can then ignore because it means we retried a batch
that was already half-applied. See KUDU-568.

No tests in this patch since we need to run against multiple TS, while our testing infra only
supports a single one. It was manually tested with YCSB against a 9 TS cluster. We'll need to
address this separately.

Change-Id: I1c339119f60bf803da2d48cd24c38880e085ec46
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5423
Reviewed-by: David Alves <david.alves@cloudera.com>
Tested-by: jenkins