[java client] Multi-TS infra for unit tests
This patch makes it so that we now start 3 tablet servers intead of just one.

TestHybridTime had to be modified because it was creating multiple tablets and
it was relying on always communicating with only one server. Now that we have
more, the results were wrong.

A new unit test was created to showcase how the client can ride over a leader
failover. This test doesn't run when configured to run against a remote cluster.

Change-Id: I95f1076b200233c9c0ce75beccd4b14863703dd7
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5563
Tested-by: jenkins
Reviewed-by: Alex Feinberg <afeinberg@cloudera.com>