Fix bug in watermark advancement when peers lag more operations behind the leader than can be sent in a single batch
In the real cluster, we were hitting a corner case where the queue was
taking the the peer's reported last_received into account wrongfully.

This makes it so that we take the last operation we sent to the peer
(and that it has successfully acked) into account, for watermark advancement
purposes, instead of the peer's reported last received operation.

Change-Id: I4fd0f63774d61ff20700d03577199914102fd33e
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5565
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins