server: set server name and return it in "wrong service" errors
This improves ServerBase so that each server is named. When an RPC call arrives
but the target service is not found, the error message now includes the name of
the server which received the call. This makes it much easier to debug cases
where the user accidentally configures the client to talk to a tablet server
instead of a master.

Change-Id: I6afabc9cf672a419f607fd6e0333905858961286
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5712
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Todd Lipcon <todd@cloudera.com>